Added Ideas.md and updated main.cpp
The ideas.md is for sharing and collaborating on ideas and also keeping track of tasks. Although I believe it can also be done through git somehow, but I don't know how yet. Also added more content in the main.cpp file and restructured it. There is some work to be done now. I'll be working on the Input function and the player class.